No more partial class Container
