+++ osu.Framework/osu.Framework.csproj
@@ -118,7 +118,6 @@
     <Compile Include="Graphics\Containers\BufferedContainerDrawNode.cs" />
     <Compile Include="Graphics\Containers\CircularContainer.cs" />
     <Compile Include="Graphics\Containers\ClickableContainer.cs" />
-    <Compile Include="Graphics\Containers\Container_AutoSize.cs" />
     <Compile Include="Graphics\Containers\FocusedOverlayContainer.cs" />
     <Compile Include="Graphics\Containers\IContainer.cs" />
     <Compile Include="Graphics\Containers\OverlayContainer.cs" />
